fixes db entities

+++ server-prod/src/main/java/com/myprojects/movieranking/repository/MoviesRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface MoviesRepository extends JpaRepository<MovieEntity,Integer> {
+public interface MoviesRepository extends JpaRepository<MovieEntity,Long> {
 
     List<MovieEntity> findByYearAbiReviewed(Integer yearAbiReviewed);
